requirements.txt
APScheduler==3.6.3
autopep8==1.7.0
cachetools==4.2.2
certifi==2022.9.24
mss==6.1.0
psutil==5.9.2
pycodestyle==2.9.1
pyperclip==1.8.2
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
toml==0.10.2
tornado==6.1
tzdata==2022.2
tzlocal==4.2
SpeechRecognition==3.9.0